using jwt api authentication for prp, moved tasks, commented code for cleanup
