+++ src/etools/applications/partners/prp_api.py
@@ -1,4 +1,3 @@
-import base64
 import json
 from typing import Dict, Iterable, NamedTuple
 
@@ -6,7 +5,10 @@ from django.conf import settings
 
 import requests
 
+from etools.applications.core.jwt_api import BaseJWTAPI
 
+
+# TODO cleanup
 class PRPPartnerResponse(NamedTuple):
     id: int
     external_id: str
@@ -14,6 +16,7 @@ class PRPPartnerResponse(NamedTuple):
     name: str
 
 
+# TODO cleanup
 class PRPPartnerUserResponse(NamedTuple):
     email: str
     title: str
@@ -23,40 +26,9 @@ class PRPPartnerUserResponse(NamedTuple):
     is_active: bool
 
 
-class PRPAPI(object):
-    def __init__(self):
-        self.url_prototype = settings.PRP_API_ENDPOINT
-        self.username = settings.PRP_API_USER
-        self.password = settings.PRP_API_PASSWORD
-        self.enabled = bool(self.url_prototype)
-
-    def _get_headers(self, data=None):
-        headers = {'Content-Type': 'application/json', 'Keep-Alive': '1800'}
-        if data:
-            headers['Content-Length'] = str(len(data))
-
-        auth_pair_str = '%s:%s' % (self.username, self.password)
-        headers['Authorization'] = 'Basic ' + \
-                                   base64.b64encode(auth_pair_str.encode()).decode()
-        return headers
-
-    def _push_request(self, data=None, timeout=None):
-        if not self.enabled:
-            return
-
-        headers = self._get_headers(data)
-
-        if data:
-            r = requests.post(url=self.url, headers=headers, json=data, verify=True, timeout=timeout)
-        else:
-            r = requests.get(url=self.url, headers=headers, verify=True, timeout=timeout)
-
-        # Any status code answer below 400 is OK
-        if r.status_code >= 400:
-            r.raise_for_status()
-
-        data = json.loads(r.text)
-        return data
+class PRPAPI(BaseJWTAPI):
+    def __init__(self, user):
+        super().__init__(user, url=settings.PRP_API_ENDPOINT)
 
     def _simple_get_request(self, timeout=None):
         if not self.enabled:
@@ -68,6 +40,7 @@ class PRPAPI(object):
 
         return json.loads(r.text)
 
+    # TODO clean up: endpoint removed in prp
     def send_partner_data(self, business_area_code: str, partner_data: Dict):
         if not self.enabled:
             return
@@ -77,6 +50,7 @@ class PRPAPI(object):
         response_data = self._push_request(data=partner_data, timeout=3000)
         return response_data
 
+    # TODO clean up: endpoint removed in prp
     def get_partners_list(self) -> Iterable[PRPPartnerResponse]:
         if not self.enabled:
             return []
@@ -94,6 +68,7 @@ class PRPAPI(object):
 
             page += 1
 
+    # TODO clean up: endpoint removed in prp
     def get_partner_staff_members(self, partner_id: int) -> Iterable[PRPPartnerUserResponse]:
         if not self.enabled:
             return []
