+++ src/etools/applications/partners/tasks.py
@@ -5,44 +5,36 @@ from django.conf import settings
 from django.contrib.auth import get_user_model
 from django.contrib.contenttypes.models import ContentType
 from django.db import connection, transaction
-from django.db.models import Prefetch
 from django.utils import timezone
 from django.utils.translation import gettext as _
 
 from celery.utils.log import get_task_logger
 from django_tenants.utils import get_tenant_model, schema_context
-from requests import HTTPError
 from unicef_snapshot.models import Activity
 from unicef_vision.exceptions import VisionException
 from unicef_vision.utils import get_data_from_insight
 
 from etools.applications.environment.notifications import send_notification_with_template
 from etools.applications.partners.models import Agreement, Intervention, PartnerOrganization
-from etools.applications.partners.prp_api import PRPAPI
-from etools.applications.partners.serializers.prp_v1 import (
-    PRPPartnerOrganizationWithStaffMembersSerializer,
-    PRPSyncUserSerializer,
-)
 from etools.applications.partners.utils import (
     copy_all_attachments,
     send_intervention_draft_notification,
     send_intervention_past_start_notification,
     send_pca_missing_notifications,
     send_pca_required_notifications,
-    sync_partner_staff_member,
 )
 from etools.applications.partners.validation.agreements import AgreementValid
 from etools.applications.partners.validation.interventions import InterventionValid
 from etools.applications.reports.models import CountryProgramme
-from etools.applications.users.models import Country, Realm, User
+from etools.applications.users.models import Country
 from etools.config.celery import app
 from etools.libraries.djangolib.utils import get_environment
-from etools.libraries.tenant_support.utils import every_country, run_on_all_tenants
+from etools.libraries.tenant_support.utils import run_on_all_tenants
 
 logger = get_task_logger(__name__)
 
 # _INTERVENTION_ENDING_SOON_DELTAS is used by intervention_notification_ending(). Notifications will be sent
-# about each interventions ending {delta} days from now.
+# about each intervention ending {delta} days from now.
 _INTERVENTION_ENDING_SOON_DELTAS = (15, 30, 60, 90)
 
 
@@ -301,42 +293,43 @@ def check_intervention_past_start():
     run_on_all_tenants(send_intervention_past_start_notification)
 
 
-@app.task
-def sync_partner_to_prp(tenant: str, partner_id: int):
-    tenant = get_tenant_model().objects.get(name=tenant)
-    connection.set_tenant(tenant)
+# TODO clean up: endpoint removed in prp
+# @app.task
+# def sync_partner_to_prp(tenant: str, partner_id: int):
+#     tenant = get_tenant_model().objects.get(name=tenant)
+#     connection.set_tenant(tenant)
+#
+#     partner = PartnerOrganization.objects.get(id=partner_id)
+#     partner_data = PRPPartnerOrganizationWithStaffMembersSerializer(instance=partner).data
+#     PRPAPI().send_partner_data(tenant.business_area_code, partner_data)
 
-    partner = PartnerOrganization.objects.get(id=partner_id)
-    partner_data = PRPPartnerOrganizationWithStaffMembersSerializer(instance=partner).data
-    PRPAPI().send_partner_data(tenant.business_area_code, partner_data)
-
-
-@app.task
-def sync_partners_staff_members_from_prp():
-    api = PRPAPI()
-
-    # remember where every particular partner is located for easier search
-    partners_tenants_mapping = {}
-    with every_country() as c:
-        for country in c:
-            connection.set_tenant(country)
-            for partner in PartnerOrganization.objects.all():
-                partners_tenants_mapping[(str(partner.id), partner.vendor_number)] = country
-
-    for partner_data in PRPAPI().get_partners_list():
-        key = (partner_data.external_id, partner_data.unicef_vendor_number)
-        partner_tenant = partners_tenants_mapping.get(key)
-        if not partner_tenant:
-            continue
-
-        connection.set_tenant(partner_tenant)
-        partner = PartnerOrganization.objects.get(
-            id=partner_data.external_id,
-            vendor_number=partner_data.unicef_vendor_number
-        )
-
-        for staff_member_data in api.get_partner_staff_members(partner_data.id):
-            sync_partner_staff_member(partner, staff_member_data)
+# TODO clean up: endpoint removed in prp
+# @app.task
+# def sync_partners_staff_members_from_prp():
+#     api = PRPAPI()
+#
+#     # remember where every particular partner is located for easier search
+#     partners_tenants_mapping = {}
+#     with every_country() as c:
+#         for country in c:
+#             connection.set_tenant(country)
+#             for partner in PartnerOrganization.objects.all():
+#                 partners_tenants_mapping[(str(partner.id), partner.vendor_number)] = country
+#
+#     for partner_data in PRPAPI().get_partners_list():
+#         key = (partner_data.external_id, partner_data.unicef_vendor_number)
+#         partner_tenant = partners_tenants_mapping.get(key)
+#         if not partner_tenant:
+#             continue
+#
+#         connection.set_tenant(partner_tenant)
+#         partner = PartnerOrganization.objects.get(
+#             id=partner_data.external_id,
+#             vendor_number=partner_data.unicef_vendor_number
+#         )
+#
+#         for staff_member_data in api.get_partner_staff_members(partner_data.id):
+#             sync_partner_staff_member(partner, staff_member_data)
 
 
 @app.task
@@ -582,30 +575,3 @@ def send_pd_to_vision(tenant_name: str, intervention_pk: int, retry_counter=0):
                              f'PD number: {intervention_pk}. Business area code: {tenant.business_area_code}')
     finally:
         connection.set_tenant(original_tenant)
-
-
-@app.task
-def sync_realms_to_prp(user_pk, last_modified_at_timestamp, retry_counter=0):
-    last_modified_instance = Realm.objects.filter(user_id=user_pk).order_by('modified').last()
-    if last_modified_instance and last_modified_instance.modified.timestamp() > last_modified_at_timestamp:
-        # there were updates to user realms. skip
-        return
-
-    user = User.objects.filter(pk=user_pk).prefetch_related(
-        Prefetch('realms', Realm.objects.filter(is_active=True).select_related('country', 'organization', 'group')),
-    ).get()
-    data = PRPSyncUserSerializer(instance=user).data
-
-    try:
-        PRPAPI().send_user_realms(data)
-    except HTTPError as ex:
-        if retry_counter < 2:
-            logger.info(f'Received {ex} from prp api. retrying')
-            sync_realms_to_prp.apply_async(
-                (user_pk, last_modified_at_timestamp),
-                {'retry_counter': retry_counter + 1},
-                eta=timezone.now() + datetime.timedelta(minutes=1 + retry_counter)
-            )
-        else:
-            logger.exception(f'Received {ex} from prp api while trying to send realms after 3 attempts. '
-                             f'User pk: {user_pk}.')
