+++ src/etools/applications/users/tasks.py
@@ -1,11 +1,18 @@
+import datetime
+
 from django.contrib.auth import get_user_model
 from django.contrib.auth.models import Group
 from django.db import connection, IntegrityError, transaction
+from django.db.models import Prefetch
+from django.utils import timezone
 
 from celery.utils.log import get_task_logger
+from requests import HTTPError
 
 from etools.applications.environment.notifications import send_notification_with_template
 from etools.applications.organizations.models import Organization
+from etools.applications.partners.prp_api import PRPAPI
+from etools.applications.partners.serializers.prp_v1 import PRPSyncUserSerializer
 from etools.applications.users.models import Country, Realm, User, UserProfile
 from etools.config.celery import app
 
@@ -213,3 +220,31 @@ def notify_user_on_realm_update(user_pk):
             template_name='users/amp/role-update',
             context=email_context
         )
+
+
+@app.task
+def sync_realms_to_prp(request_user_pk, user_pk, last_modified_at_timestamp, retry_counter=0):
+    last_modified_instance = Realm.objects.filter(user_id=user_pk).order_by('modified').last()
+    if last_modified_instance and last_modified_instance.modified.timestamp() > last_modified_at_timestamp:
+        # there were updates to user realms. skip
+        return
+    request_user = User.objects.get(pk=request_user_pk)
+
+    user = User.objects.filter(pk=user_pk).prefetch_related(
+        Prefetch('realms', Realm.objects.filter(is_active=True).select_related('country', 'organization', 'group')),
+    ).get()
+    data = PRPSyncUserSerializer(instance=user).data
+
+    try:
+        PRPAPI(request_user).send_user_realms(data)
+    except HTTPError as ex:
+        if retry_counter < 2:
+            logger.info(f'Received {ex} from prp api. retrying')
+            sync_realms_to_prp.apply_async(
+                (request_user_pk, user_pk, last_modified_at_timestamp),
+                {'retry_counter': retry_counter + 1},
+                eta=timezone.now() + datetime.timedelta(minutes=1 + retry_counter)
+            )
+        else:
+            logger.exception(f'Received {ex} from prp api while trying to send realms after 3 attempts. '
+                             f'User pk: {user_pk}.')
